+++ packages/core/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.6.2",
+    "csstype": "^3.0.5",
     "emotion": "^10.0.17",
     "@contentful/f36-tokens": "^4.0.0"
   },
