+++ packages/core/src/Grid/Grid.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import { axe } from '../../utils/axeHelper';
+import { axe } from '../../../../scripts/test/axeHelper';
 import { Grid } from './Grid';
 
 it('renders the component', () => {
