+++ packages/core/src/Primitive/index.ts
@@ -1,0 +1,6 @@
+export { Primitive } from './Primitive';
+export type {
+  PrimitiveProps,
+  PolymorphicComponentProps,
+  PolymorphicComponent,
+} from './Primitive';
