--- packages/forma-36-react-components/src/components/Grid/Grid.css
@@ -1,11 +1,0 @@
-@import 'settings/colors';
-@import 'settings/typography';
-@import 'settings/transitions';
-
-.Grid {
-  display: grid;
-}
-
-.Grid__inline {
-  display: inline-grid;
-}
